Move tools to a separate file
